check connecting before sending SIGINT

+++ reader.py
@@ -3,7 +3,7 @@ import nfc
 
 def sig_handler(signo, frame):
     sig_handler.running ^= True
-    if not(sig_handler.running):
+    if not(sig_handler.running) and read_id.connecting:
         os.kill(os.getpid(), signal.SIGINT)
 
 sig_handler.running = True
@@ -25,12 +25,16 @@ def read_id():
     if sig_handler.running:
         clf = nfc.ContactlessFrontend('usb')
         try:
+            read_id.connecting = True
             clf.connect(rdwr={'on-connect':on_connect})
         finally:
+            read_id.connecting = False
             stdout_json({'event':'touchend'})
             clf.close()
     else:
         time.sleep(0.1)
+
+read_id.connecting = False
 
 if __name__ == '__main__':
     while True:
